requirements.txt
# Web Framework
Flask>=3.0.0
Flask-Login>=0.6.0
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.2.0
Flask-Migrate>=4.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-Talisman>=1.0.0

# Forms & Validation
WTForms>=3.0.0
email-validator>=2.1.0

# Security & Auth
itsdangerous>=2.1.0
PyJWT>=2.8.0
Werkzeug>=3.0.0

# Data & API
requests>=2.31.0
python-dotenv>=1.0.0

# Machine Learning & Embeddings
numpy>=1.24.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Production Server
gunicorn>=21.2.0
